V1.0.0.4 And V1.0.0.5 Uploaded
in V1.0.0.5, "Import" Function is More Convenient. You can declare a functiion called "ImprotExcute" and the code inside will be ran as global code while the model is imported. It also provides you a Variant Object called "ImportInfos" and you can get "ModelFullPath" and "MainScriptFullPath" In ImprotExcute Function as well.